Implamented tracker response parsing with dictionary peer model

+++ BitTorrentEdu/DTOs/Peer.cs
@@ -9,7 +9,7 @@ namespace BitTorrentEdu.DTOs
 {
    public class Peer
     {
-        public Peer(IPAddress ip, int port, string id = null)
+        public Peer(IPAddress ip, long port, string id = null)
         {
             Id = id;
             Ip = ip;
@@ -18,6 +18,6 @@ namespace BitTorrentEdu.DTOs
 
         public string Id { get; private set; }
         public IPAddress Ip { get; private set; }
-        public int Port { get; private set; }
+        public long Port { get; private set; }
     }
 }
